requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2
chromadb>=0.4.22
pypdf>=3.17.4
docx2txt>=0.8
numpy>=1.24.3
scikit-learn>=1.3.0
aiohttp>=3.9.1
requests>=2.31.0.0